BAP-17535: Decrease the number of EXISTS calls to Redis (#20020)

+++ src/Oro/Bundle/WebsiteBundle/Provider/CacheableWebsiteProvider.php
@@ -79,6 +79,10 @@ class CacheableWebsiteProvider extends BaseWebsiteProvider
      */
     public function clearCache()
     {
+        if (!$this->hasCache()) {
+            return;
+        }
+
         $this->cacheProvider->delete(self::WEBSITE_IDS_CACHE_KEY);
     }
 }
